pyproject.toml
[tool.poetry]
name = "sevenrealms-bot"
version = "0.1.0"
description = ""
authors = ["Zeithrold <zeithrold@gmail.com>"]
readme = "README.md"
packages = [{ include = "sevenrealms_bot" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
nonebot2 = "^2.1.0"
uvicorn = "^0.23.2"
ghapi = "^1.0.4"
pony = "^0.7.16"
pymysql = "^1.0.2"
nonebot-adapter-onebot = "^2.2.0"
nonebot-adapter-github = "^0.3.0"
nonebot-plugin-status = "^0.6.0"
fastapi = "^0.103.1"
websockets = "^10.4"
nonebot-plugin-apscheduler = "^0.2.0"
aliyun-python-sdk-alinlp = "^1.0.16"
oss2 = "^2.16.0"
pandas = "^2.1.0"
py7zr = "^0.20.6"
nonebot-plugin-simplemusic = "^0.1.7"
nonebot-plugin-miragetank = "^0.1.3"
nonebot-plugin-datastore = "^1.1.1"
nonebot-plugin-groupmate-waifu = "^1.3.3"
nonebot-plugin-setu-collection = "^0.1.4"
nonebot-plugin-jrrp2 = "^1.6"
nonebot-plugin-pjsk = "^0.2.9"
nb-cli = ">=1.0.0,<1.1.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
asyncmy = "^0.2.9"
tqdm = "^4.66.1"


[tool.nonebot]
plugins = [
    "nonebot_plugin_status",
    "nonebot_plugin_apscheduler",
    "nonebot_plugin_simplemusic",
    "nonebot_plugin_miragetank",
    "nonebot_plugin_datastore",
    "nonebot_plugin_groupmate_waifu",
    "nonebot_plugin_setu_collection",
    "nonebot_plugin_jrrp2",
    "nonebot_plugin_pjsk"
]
plugin_dirs = ["sevenrealms_bot/plugins"]

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
